Move iterator_pair to the container module
