+++ modules/container/CMakeSources.cmake
@@ -5,6 +5,7 @@ set(MODULE_PUBLIC_HEADERS
         "${MODULE_PATH}/array_ops.hpp"
         "${MODULE_PATH}/container_view.hpp"
         "${MODULE_PATH}/contextful_container.hpp"
+        "${MODULE_PATH}/iterator_pair.hpp"
         "${MODULE_PATH}/vect.hpp"
         )
